requirements.txt
# Requirements for App
streamlit
specklepy
pandas
streamlit-aggrid
streamlit-on-Hover-tabs
plotly-express
numpy